requirements.txt
asgiref==3.11.1
attrs==26.1.0
dj-database-url==3.1.2
Django==6.0.1
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
drf-spectacular==0.29.0
gunicorn==25.3.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Markdown==3.10.2
packaging==26.0
psycopg2-binary==2.9.11
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
whitenoise==6.12.0
